change Sign uId field type from int to long

+++ app/src/main/java/dz/deepwork/gamemn02/presenceregister/data/signs/Sign.java
@@ -4,7 +4,7 @@ import dz.deepwork.gamemn02.presenceregister.data.signins.SignIn;
 
 public class Sign {
 
-    public int uId;
+    public long uId;
     public String memberName;
     public long signInTime;
     public long signOutTime;
